+++ changelog.md
@@ -269,4 +269,5 @@ Breaking changes:
 * Throw specific error if try to create child of non-existent parent (closes #21)
 * Support sequelize's `fields` option
 * Refactor
+* Tests refactor
 * Increase tests timeout for coveralls